requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT==0.3.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mysqlclient==2.1.0
packaging==21.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.0
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.32
toml==0.10.2
Werkzeug==2.0.3
